requirements.txt
selenium
python-dotenv
# chromedriver-binary-auto